Move tests and add lint/coverage config for dataentryconfig
