+++ .testcoverage.yml
@@ -30,7 +30,9 @@ override:
   - path: ^internal/importer$
     threshold: 65
   - path: ^internal/dataentry$
-    threshold: 60
+    threshold: 55
+  - path: ^internal/dataentryconfig$
+    threshold: 70
 
 # Exclude packages that are not meaningful to coverage-check
 exclude:
